Removed old scroller and simple-table
