--- src/client/components/scroller/scroller.mocha.tsx
@@ -1,25 +1,0 @@
-import { expect } from 'chai';
-import * as sinon from 'sinon';
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import * as TestUtils from 'react-addons-test-utils';
-
-import '../../utils/test-utils/index';
-
-import { $, Expression } from 'plywood';
-import { Scroller } from './scroller';
-
-describe.skip('Scroller', () => {
-  it('adds the correct class', () => {
-    var renderedComponent = TestUtils.renderIntoDocument(
-      <Scroller
-        ref="base"
-        onScroll={() => {}}
-      />
-    );
-
-    expect(TestUtils.isCompositeComponent(renderedComponent), 'should be composite').to.equal(true);
-    expect((ReactDOM.findDOMNode(renderedComponent) as any).className, 'should contain class').to.contain('scroll-container');
-  });
-
-});
