--- src/client/components/scroller/simple-scroller.scss
@@ -1,34 +1,0 @@
-.simple-scroller {
-  overflow: hidden;
-  cursor: pointer;
-
-  > * {
-    position: absolute;
-    overflow: hidden;
-    white-space: nowrap;
-  }
-
-  > .event-container {
-    overflow: auto;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-  }
-
-  .left-shadow {
-    box-shadow: 1px 0 5px 0 rgba(0, 0, 0, 0.2);
-  }
-
-  .top-shadow {
-    box-shadow: 3px 0 5px 1px rgba(0, 0, 0, 0.2);
-  }
-
-  .right-shadow {
-    box-shadow: 4px 0 4px 5px hsla(0, 0, 0, 0.2);
-  }
-
-  .bottom-shadow {
-    box-shadow: 0 0 2px 1px rgba(0, 0, 0, 0.2);
-  }
-}
