--- src/client/components/simple-table/simple-table.mocha.tsx
@@ -1,30 +1,0 @@
-import { expect } from 'chai';
-import * as sinon from 'sinon';
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import * as TestUtils from 'react-addons-test-utils';
-
-import '../../utils/test-utils/index';
-
-import { $, Expression } from 'plywood';
-import { SimpleTable } from './simple-table';
-
-describe('SimpleTable', () => {
-  it('adds the correct class', () => {
-    var renderedComponent = TestUtils.renderIntoDocument(
-      <SimpleTable
-        scrollLeft={2}
-        scrollTop={3}
-        headerColumns={null}
-        rows={null}
-        rowWidth={100}
-        dataLength={30}
-        rowHeight={30}
-      />
-    );
-
-    expect(TestUtils.isCompositeComponent(renderedComponent), 'should be composite').to.equal(true);
-    expect((ReactDOM.findDOMNode(renderedComponent) as any).className, 'should contain class').to.contain('simple-table');
-  });
-
-});
