--- src/client/components/simple-table/simple-table.scss
@@ -1,40 +1,0 @@
-@import '../../imports';
-
-.simple-table {
-  @include pin-full;
-
-  .pre-body {
-    @include pin-full;
-  }
-
-  .header-cont {
-    @include pin-full;
-    overflow: hidden;
-
-    .header {
-      border-bottom: 1px solid $border-extra-light;
-      white-space: nowrap;
-    }
-  }
-
-  .body-cont {
-    overflow: hidden;
-
-    .row {
-      position: absolute;
-      border-bottom: 1px solid $border-extra-light;
-      left: 0;
-      white-space: nowrap;
-    }
-  }
-
-  .horizontal-scroll-shadow {
-    position: absolute;
-    top: 0;
-    box-shadow: 0 4px 4px -3px hsla(0, 0, 0, 0.2);
-  }
-
-  .post-body {
-    @include pin-full;
-  }
-}
