+++ src/client/visualizations/table/table.scss
@@ -61,7 +61,7 @@ $header-padding-top: 12px;
     }
   }
 
-  .simple-scroller {
+  .scroller {
     position: absolute;
     top: 0;
     bottom: 0;
